useGenericTextFieldInContext hook
