+++ front/src/modules/ui/table/editable-cell/type/components/GenericEditableTextCell.tsx
@@ -1,10 +1,7 @@
-import { useRecoilValue } from 'recoil';
-
 import { TextDisplay } from '@/ui/content-display/components/TextDisplay';
 import type { ViewFieldTextMetadata } from '@/ui/editable-field/types/ViewField';
 import { EditableCell } from '@/ui/table/editable-cell/components/EditableCell';
-import { useCurrentRowEntityId } from '@/ui/table/hooks/useCurrentEntityId';
-import { tableEntityFieldFamilySelector } from '@/ui/table/states/selectors/tableEntityFieldFamilySelector';
+import { useGenericTextFieldInContext } from '@/ui/table/hooks/useGenericTextFieldInContext';
 
 import type { ColumnDefinition } from '../../../types/ColumnDefinition';
 
@@ -19,14 +16,7 @@ export function GenericEditableTextCell({
   columnDefinition,
   editModeHorizontalAlign,
 }: OwnProps) {
-  const currentRowEntityId = useCurrentRowEntityId();
-
-  const fieldValue = useRecoilValue<string>(
-    tableEntityFieldFamilySelector({
-      entityId: currentRowEntityId ?? '',
-      fieldName: columnDefinition.metadata.fieldName,
-    }),
-  );
+  const { fieldValue } = useGenericTextFieldInContext();
 
   return (
     <EditableCell
