+++ front/src/modules/ui/table/editable-cell/type/components/GenericEditableTextCellEditMode.tsx
@@ -1,10 +1,6 @@
-import { useRecoilState } from 'recoil';
-
 import type { ViewFieldTextMetadata } from '@/ui/editable-field/types/ViewField';
 import { useCellInputEventHandlers } from '@/ui/table/hooks/useCellInputEventHandlers';
-import { useCurrentRowEntityId } from '@/ui/table/hooks/useCurrentEntityId';
-import { useUpdateEntityField } from '@/ui/table/hooks/useUpdateEntityField';
-import { tableEntityFieldFamilySelector } from '@/ui/table/states/selectors/tableEntityFieldFamilySelector';
+import { useGenericTextFieldInContext } from '@/ui/table/hooks/useGenericTextFieldInContext';
 import { TableHotkeyScope } from '@/ui/table/types/TableHotkeyScope';
 
 import { TextInput } from '../../../../input/components/TextInput';
@@ -17,26 +13,15 @@ type OwnProps = {
 export function GenericEditableTextCellEditMode({
   columnDefinition,
 }: OwnProps) {
-  const currentRowEntityId = useCurrentRowEntityId();
-
-  // TODO: we could use a hook that would return the field value with the right type
-  const [fieldValue, setFieldValue] = useRecoilState<string>(
-    tableEntityFieldFamilySelector({
-      entityId: currentRowEntityId ?? '',
-      fieldName: columnDefinition.metadata.fieldName,
-    }),
-  );
-
-  const updateField = useUpdateEntityField();
+  const { fieldValue, setFieldValue, updateTextField } =
+    useGenericTextFieldInContext();
 
   function handleSubmit(newText: string) {
     if (newText === fieldValue) return;
 
     setFieldValue(newText);
 
-    if (currentRowEntityId && updateField) {
-      updateField(currentRowEntityId, columnDefinition, newText);
-    }
+    updateTextField(newText);
   }
 
   const {
